Change formatting of location indicator
